Use polyfilled AbortController for abortable fetch() detection if necessary
If a browser supports AbortController/AbortSignal already then we want to use
these for the "feature detection" where we check whether the provided "fetch"
can handle abort or not, because otherwise there might be a risk that the
browser will not leave the abort signal intact while passing through the Request()
constructor for the wrong reasons (i.e. because the browser detects that the
passed signal is not a proper AbortSignal and not because the fetch implementation
lacks abort support -- which is the thing we _want_ to detect).

However, when polyfilling abort functionality on top of "node-fetch" in Node.js
then it's not likely that global.AbortController is defined because it might be
imported as "const { AbortController, abortableFetch } = require(..." rather than
"global.AbortController = require(...).AbortController" so it's not safe to
assume that "abortable fetch is not supported whenever AbortController is not
defined in the global scope".