Export a module that polyfills AbortController/AbortSignal without patching fetch
This is useful if you are using a fetch polyfill that implements abortable fetch
but that does not itself bundle a polyfill for AbortController (the AbortController
class is not specific to fetch, it might be used in other DOM APIs in the future).

For this use case you would install the AbortController polyfill first and then
the fetch polyfill (instead of the other way around):

  import 'abortcontroller-polyfill/dist/abortcontroller-polyfill-only'
  import '... fetch polyfill package ...'

This commit also exports a module that DOES patch fetch, in addition to also
installing the abortcontroller-polyfill. This is exactly what the default package
export does today, i.e. the following two lines are equivalent:

  import 'abortcontroller-polyfill'
  import 'abortcontroller-polyfill/dist/polyfill-patch-fetch'

Long term, if whatwg-fetch, unfetch and node-fetch implement abortable fetch
it makes sense to release a new major version of abortcontroller-polyfill where
the package default export is the "abortcontroller-polyfill-only" module instead.