Install polyfill on webkit-based browsers even if window.AbortController exists
The native AbortController is just a stub that doesn't actually abort. This affects
Safari 12.0.1 on Mac OS and also the iOS version. Chrome and Firefox on iOS are also
affected since they also share the same webkit rending engine.

A working implementation of window.AbortController in these browsers is tracked by:
https://bugs.webkit.org/show_bug.cgi?id=174980