Detect and polyfill over present-but-broken native AbortController in Safari 12.0.x, 12.1.x specifically but not >= 12.2
The native AbortController in Safari 12.2 and above works (starting with Safari Technology Preview Release 74) so the
polyfill should not install for these versions.

The latest versions of Chrome/Firefox for IOS still have the "present but broken" native AbortController so always
polyfill over their native implementations for now (the symptom otherwise is that cancelling never works, the HTTP
call happens even if the .abort() happens before the fetch() and the fetch promise is not rejected.

This commit fixes github issue #26