Support force installing polyfill on top of browser with native AbortController
Will force install polyfill if:

  window.__FORCE_INSTALL_ABORTCONTROLLER_POLYFILL == true

is set (this is mostly for testing purposes). When running in this mode, it is
important to never pass a fake AbortSignal object to the native implementations
of "Request" and "fetch" because these will respond by throwing TypeError in
many browsers (like Chrome for example).